requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
streamlit>=1.33.0
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
requests>=2.31.0 